requirements_full.txt
altair==5.5.0
attrs==25.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
docopt==0.6.2
filelock==3.18.0
Flask==3.1.0
fsspec==2025.3.2
gitdb==4.0.12
GitPython==3.1.44
gTTS==2.5.4
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
Js2Py==0.74
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
keyboard==0.13.5
llvmlite==0.44.0
MarkupSafe==3.0.2
more-itertools==10.7.0
mpmath==1.3.0
narwhals==1.36.0
networkx==3.4.2
numba==0.61.2
numpy==2.2.5
openai-whisper @ git+https://github.com/openai/whisper.git@517a43ecd132a2089d85f4ebc044728a71d49f6e
packaging==24.2
pandas==2.2.3
pillow==11.2.1
pipwin==0.5.2
protobuf==5.29.4
pyarrow==19.0.1
pycparser==2.22
pydeck==0.9.1
pydub==0.25.1
pyjsparser==2.7.1
PyPrind==2.11.3
pySmartDL==1.3.4
python-dateutil==2.9.0.post0
pytz==2025.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rpds-py==0.24.0
scipy==1.15.2
setuptools==79.0.0
six==1.17.0
smmap==5.0.2
sounddevice==0.5.1
soundfile==0.13.1
soupsieve==2.7
streamlit==1.44.1
streamlit-audiorecorder==0.0.6
sympy==1.13.1
tenacity==9.1.2
tiktoken==0.9.0
toml==0.10.2
torch==2.6.0
tornado==6.4.2
tqdm==4.67.1
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.4.0
watchdog==6.0.0
Werkzeug==3.1.3
wheel==0.45.1
